Removed warning about signed/unsigned int comparison

+++ crts.cpp
@@ -1752,7 +1752,7 @@ int main(int argc, char ** argv)
                         if (verbose) printf("\n\nGenerating data that will go in frame...\n");
                         for (i=0; i<8; i++)
                             header[i] = i & 0xff;
-                        for (i=0; i<ce.payloadLen; i++)
+                        for (i=0; i<(signed int) ce.payloadLen; i++)
                             payload[i] = i & 0xff;
                         // Include frame number in header information
                         * header_f = ce.frameNumber;
@@ -1805,7 +1805,7 @@ int main(int argc, char ** argv)
                         if (verbose) printf("\n\nGenerating data that will go in frame...\n");
                         for (i=0; i<8; i++)
                             header[i] = i & 0xff;
-                        for (i=0; i<ce.payloadLen; i++)
+                        for (i=0; i<(signed int)ce.payloadLen; i++)
                             payload[i] = i & 0xff;
                         // Include frame number in header information
                         * header_f = ce.frameNumber;
